// src/model/types/EntrySectionType.ts
enum EntrySectionType {
  Function = 'Function',
  FamilyAndDomains = 'Family & Domains',
  Expression = 'Expression',
  Interaction = 'Interaction',
  Miscellaneous = 'Miscellaneous',
  NamesAndTaxonomy = 'Names & Taxonomy',
  PathologyAndBioTech = 'Pathology & BioTech',
  ProteinProcessing = 'Protein Processing',
  Sequence = 'Sequence',
  Structure = 'Structure',
  SubCellularLocation = 'Subcellular Location',
}

export default EntrySectionType;